% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxVisiumData.R
\name{TENxVisiumData}
\alias{TENxVisiumData}
\title{TENxVisiumData}
\usage{
TENxVisiumData(
  dataset = c("HumanBreastCancer", "HumanHeart", "HumanLymphNode",
    "MouseKidneyCoronal", "MouseBrainCoronal", "MouseBrainSagittalPosterior",
    "MouseBrainSagittalAnterior")
)
}
\arguments{
\item{dataset}{character string specifying which dataset to load.}
}
\description{
Function to load 10X Genomics Visium spatial gene expression datasets
from the \code{TENxVisiumData} \code{ExperimentHub} package.
}
\examples{
(spe <- TENxVisiumData("HumanHeart"))

}
